跨域配置

+++ src/main/java/dictation/word/service/i/user/TokenService.java
@@ -33,9 +33,9 @@ public interface TokenService {
      *
      * @return userId
      */
-    default long getCurrentUserId() {
+    default int getCurrentUserId() {
         final JSONObject token = getCurrentToken();
-        return token.getLong("uid");
+        return token.getInteger("uid");
     }
 
     /**
